+++ src/main/java/com/example/backendmainserver/PowerData/domain/PowerDataList.java
@@ -1,5 +1,7 @@
-package com.example.backendmainserver.websocket.presentation.dto;
+package com.example.backendmainserver.PowerData.domain;
 
+import com.example.backendmainserver.PowerData.domain.PowerData;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,6 +9,7 @@ import java.util.List;
 
 @Getter
 @Setter
+@Builder
 public class PowerDataList {
     private List<PowerData> powerDataList;
